admin: factor out cluster config routes
